+++ tests/test_bar_plugin.py
@@ -290,6 +290,7 @@ class TestBarPlugin:
 
         # We need to parse the version to account for alpha release,
         # like Python `3.12.0a4`.
-        assert parse_version(process.stdout.split()[1]) >= parse_version(
+        python_version = process.stdout.split()[1]
+        assert parse_version(python_version) >= parse_version(
             ".".join(str(i) for i in bar_plugin.PYTHON_MIN_VERSION),
-        )
+        ), f"{python_version} >= {bar_plugin.PYTHON_MIN_VERSION}"